requirements.txt
albumentations==1.1.0
lpips==0.1.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.5.62
openTSNE==0.6.1
pandas==1.1.5
Pillow==8.4.0
pytorch-metric-learning==1.1.0
pytorch_msssim==0.2.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.4.1
toml==0.10.2
torch==1.10.0
torchmetrics==0.7.0
torchvision
tqdm==4.62.3
umap-learn==0.5.2